pyproject.toml
[project]
name = "awesome-python"
version = "0.1.0"
description = "An opinionated list of awesome Python frameworks, libraries, software and resources."
requires-python = ">=3.13"
dependencies = [
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "markdown==3.10.2",
    "markdown-it-py==4.0.0",
]

[dependency-groups]
dev = [
    "pytest==9.0.2",
    "ruff==0.15.6",
]

[tool.pytest.ini_options]
testpaths = ["website/tests"]

[tool.ruff]
target-version = "py313"
line-length = 100
